# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "IBKR REST Bridge API",
    "description": "RESTful API bridge for Interactive Brokers Client Portal Gateway",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "Basic authentication for API access"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/orders": {
      "get": {
        "tags": [
          "Orders"
        ],
        "description": "Get all orders for the account",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "orders": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "orderId": {
                            "type": "string",
                            "example": "1234567890"
                          },
                          "accountId": {
                            "type": "string",
                            "example": "U1234567"
                          },
                          "instrument": {
                            "type": "object",
                            "properties": {
                              "conid": {
                                "type": "integer",
                                "example": 265598
                              },
                              "symbol": {
                                "type": "string",
                                "example": "AAPL"
                              },
                              "type": {
                                "type": "string",
                                "enum": [
                                  "stock",
                                  "option",
                                  "future",
                                  "forex",
                                  "other"
                                ],
                                "example": "stock"
                              },
                              "exchange": {
                                "type": "string",
                                "example": "NASDAQ"
                              },
                              "currency": {
                                "type": "string",
                                "example": "USD"
                              }
                            },
                            "required": [
                              "conid",
                              "symbol",
                              "type"
                            ]
                          },
                          "side": {
                            "type": "string",
                            "enum": [
                              "buy",
                              "sell"
                            ],
                            "example": "buy"
                          },
                          "type": {
                            "type": "string",
                            "enum": [
                              "market",
                              "limit"
                            ],
                            "example": "limit"
                          },
                          "quantity": {
                            "type": "number",
                            "example": 100
                          },
                          "limitPrice": {
                            "type": "number",
                            "example": 150
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "pending",
                              "submitted",
                              "filled",
                              "partially_filled",
                              "cancelled",
                              "rejected"
                            ],
                            "example": "submitted"
                          },
                          "filledQuantity": {
                            "type": "number",
                            "example": 0
                          },
                          "avgFillPrice": {
                            "type": "number",
                            "example": 150.25
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "orderId",
                          "accountId",
                          "instrument",
                          "side",
                          "type",
                          "quantity",
                          "status",
                          "filledQuantity",
                          "createdAt",
                          "updatedAt"
                        ]
                      }
                    }
                  },
                  "required": [
                    "orders"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Orders"
        ],
        "description": "Place a new order",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "conid": {
                    "type": "integer",
                    "description": "Contract ID of the instrument",
                    "example": 265598
                  },
                  "side": {
                    "type": "string",
                    "enum": [
                      "buy",
                      "sell"
                    ],
                    "example": "buy"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "market",
                      "limit"
                    ],
                    "example": "limit"
                  },
                  "quantity": {
                    "type": "number",
                    "minimum": 1,
                    "example": 100
                  },
                  "limitPrice": {
                    "type": "number",
                    "description": "Required for limit orders",
                    "example": 150
                  }
                },
                "required": [
                  "conid",
                  "side",
                  "type",
                  "quantity"
                ]
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "orderId": {
                      "type": "string",
                      "example": "1234567890"
                    },
                    "accountId": {
                      "type": "string",
                      "example": "U1234567"
                    },
                    "instrument": {
                      "type": "object",
                      "properties": {
                        "conid": {
                          "type": "integer",
                          "example": 265598
                        },
                        "symbol": {
                          "type": "string",
                          "example": "AAPL"
                        },
                        "type": {
                          "type": "string",
                          "enum": [
                            "stock",
                            "option",
                            "future",
                            "forex",
                            "other"
                          ],
                          "example": "stock"
                        },
                        "exchange": {
                          "type": "string",
                          "example": "NASDAQ"
                        },
                        "currency": {
                          "type": "string",
                          "example": "USD"
                        }
                      },
                      "required": [
                        "conid",
                        "symbol",
                        "type"
                      ]
                    },
                    "side": {
                      "type": "string",
                      "enum": [
                        "buy",
                        "sell"
                      ],
                      "example": "buy"
                    },
                    "type": {
                      "type": "string",
                      "enum": [
                        "market",
                        "limit"
                      ],
                      "example": "limit"
                    },
                    "quantity": {
                      "type": "number",
                      "example": 100
                    },
                    "limitPrice": {
                      "type": "number",
                      "example": 150
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "pending",
                        "submitted",
                        "filled",
                        "partially_filled",
                        "cancelled",
                        "rejected"
                      ],
                      "example": "submitted"
                    },
                    "filledQuantity": {
                      "type": "number",
                      "example": 0
                    },
                    "avgFillPrice": {
                      "type": "number",
                      "example": 150.25
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "orderId",
                    "accountId",
                    "instrument",
                    "side",
                    "type",
                    "quantity",
                    "status",
                    "filledQuantity",
                    "createdAt",
                    "updatedAt"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/{orderId}": {
      "put": {
        "tags": [
          "Orders"
        ],
        "description": "Modify an existing order",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "quantity": {
                    "type": "number",
                    "minimum": 1,
                    "example": 50
                  },
                  "limitPrice": {
                    "type": "number",
                    "example": 155
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "orderId",
            "required": true,
            "description": "Order ID to modify"
          }
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "orderId": {
                      "type": "string",
                      "example": "1234567890"
                    },
                    "accountId": {
                      "type": "string",
                      "example": "U1234567"
                    },
                    "instrument": {
                      "type": "object",
                      "properties": {
                        "conid": {
                          "type": "integer",
                          "example": 265598
                        },
                        "symbol": {
                          "type": "string",
                          "example": "AAPL"
                        },
                        "type": {
                          "type": "string",
                          "enum": [
                            "stock",
                            "option",
                            "future",
                            "forex",
                            "other"
                          ],
                          "example": "stock"
                        },
                        "exchange": {
                          "type": "string",
                          "example": "NASDAQ"
                        },
                        "currency": {
                          "type": "string",
                          "example": "USD"
                        }
                      },
                      "required": [
                        "conid",
                        "symbol",
                        "type"
                      ]
                    },
                    "side": {
                      "type": "string",
                      "enum": [
                        "buy",
                        "sell"
                      ],
                      "example": "buy"
                    },
                    "type": {
                      "type": "string",
                      "enum": [
                        "market",
                        "limit"
                      ],
                      "example": "limit"
                    },
                    "quantity": {
                      "type": "number",
                      "example": 100
                    },
                    "limitPrice": {
                      "type": "number",
                      "example": 150
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "pending",
                        "submitted",
                        "filled",
                        "partially_filled",
                        "cancelled",
                        "rejected"
                      ],
                      "example": "submitted"
                    },
                    "filledQuantity": {
                      "type": "number",
                      "example": 0
                    },
                    "avgFillPrice": {
                      "type": "number",
                      "example": 150.25
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "orderId",
                    "accountId",
                    "instrument",
                    "side",
                    "type",
                    "quantity",
                    "status",
                    "filledQuantity",
                    "createdAt",
                    "updatedAt"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Orders"
        ],
        "description": "Cancel an order",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "orderId",
            "required": true,
            "description": "Order ID to cancel"
          }
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Order cancelled successfully"
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
}